Added jQuery file and updated index.html
Updated the index.html file such that it uses the specific version of
jQuery that we are targeting for this project (2.0.3).
Likewise updated the index.html file such that the scripts are loaded at
the end of the body of the html file. This is better practice since the
scripts won't load until the html has all but finished loading. This is
especially important with regards to jQuery, since it won't run until we
know that the document is ready (HTML completely loaded).